requirements[1].txt
flask~=3.1.1
python-dotenv~=1.1.1
requests~=2.32.3

twilio~=9.7.0